[APIBREAK] Split the Search class into two classes
Separate methods about HTML export and methodes about RDF DCat
