+++ pgmetadata/lib/DCatSupportException.php
@@ -1,0 +1,15 @@
+<?php
+/**
+ * @author    Laurent Jouanneau
+ * @copyright 2022 3liz
+ *
+ * @see      https://3liz.com
+ *
+ * @license    Mozilla Public Licence
+ */
+
+namespace PgMetadata;
+
+class DCatSupportException extends \Exception
+{
+}
